kpiMachineLearningModel/requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
click==6.7
Flask==0.12.2
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysql-connector==2.1.6
numpy==1.14.2
protobuf==3.5.2.post1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
Werkzeug==0.14.1
